Remove system for detecting blocking calls that never returned.
The current system attempted to use 2-steps:
1) Record when we entered the call.
2) Record when we are done.

However, this doesn't work. Blocking calls don't always cleanly do #1.
So we see no blocking event, and assume we desynchronized.

This could technically be a desync, as it means the program is racing:
usually between ending execution. (Our tests/examples do this as
well).
But it is not very useful since we end up being too conservative and
dishing out a bunch of desync errors.

Maybe in the future we would like a "conservative mode" in case people
do care about this type of racy events.